requirements.txt
Pillow>=10.2.0
reportlab>=4.1.0
click>=8.0.0
tqdm>=4.65.0 